@@ -1,6 +1,6 @@
 {
     "name": "@enterwell/react-form-validation",
-    "version": "1.2.8",
+    "version": "1.3.0",
     "description": "Form validation functionality for React components",
     "license": "MIT",
     "repository": {
@@ -17,7 +17,7 @@
     },
     "scripts": {
         "build": "babel src/ -d dist",
-        "test": "cross-env NODE_ENV=test cypress run",
+        "test": "cross-env NODE_ENV=test cypress run --component",
         "test:open": "cross-env NODE_ENV=test cypress open"
     },
     "types": "./types/index.d.ts",
@@ -34,19 +34,22 @@
     ],
     "dependencies": {},
     "devDependencies": {
-        "@babel/cli": "^7.10.1",
-        "@babel/core": "^7.10.1",
-        "@babel/preset-env": "^7.10.1",
-        "@babel/preset-react": "^7.12.1",
-        "@cypress/code-coverage": "^3.8.2",
-        "babel-loader": "^8.1.0",
-        "babel-plugin-istanbul": "^6.0.0",
-        "cross-env": "^7.0.3",
-        "cypress": "^5.5.0",
-        "cypress-react-unit-test": "^4.16.7",
+        "@babel/cli": "7.23.0",
+        "@babel/core": "7.23.0",
+        "@babel/preset-env": "7.22.20",
+        "@babel/preset-react": "7.22.15",
+        "@cypress/code-coverage": "3.12.1",
+        "@cypress/react18": "^2.0.0",
+        "@types/react-dom": "18",
+        "@vitejs/plugin-react": "^4.1.0",
+        "babel-loader": "9.1.3",
+        "babel-plugin-istanbul": "6.1.1",
+        "cross-env": "7.0.3",
+        "cypress": "13.3.0",
         "react": "^18",
         "react-dom": "^18",
-        "webpack": "^4.44.2"
+        "vite": "4.4.9",
+        "webpack": "^5.89.0"
     },
     "peerDependencies": {
         "react": "^16.13.1 || ^17 || ^18",
