@@ -1,12 +1,14 @@
-import React from 'react';
+import React, { useEffect } from 'react';
+import { mount } from '@cypress/react18';
+import { isValidEmail } from '../../src';
 import EmailComponentWithFields from '../testComponents/EmailComponentWithFields.jsx';
 import EmailComponentWithLocalValidation from '../testComponents/EmailComponentWithLocalValidation.jsx';
 import EmailComponentWithRemoteValidation from '../testComponents/EmailComponentWithRemoteValidation.jsx';
 import EmailComponentWithSubmit from '../testComponents/EmailComponentWithSubmit.jsx';
-const { mount } = require("cypress-react-unit-test/")
+import { submitForm } from '../../src/formUtils.js';
+import { useValidation } from '../../src/useValidation.js';
 
-const emailValidation = (value) => /^[a-zA-Z0-9.!#$%&’*+/=?^_`{|}~-]+@[a-zA-Z0-9-]+(?:\.[a-zA-Z0-9-]+)*$/.test(value);
-const emailValidationRemote = (value) => new Promise((resolve) => setTimeout(() => resolve(/^[a-zA-Z0-9.!#$%&’*+/=?^_`{|}~-]+@[a-zA-Z0-9-]+(?:\.[a-zA-Z0-9-]+)*$/.test(value)), 2000));
+const emailValidationRemote = (value) => new Promise((resolve) => setTimeout(() => resolve(isValidEmail(value)), 2000));
 const testEmailValid = "matej.radovix@enterwell.net";
 const testEmailInvalid = "matej.radovix";
 
@@ -20,7 +22,7 @@ describe('useValidation', () => {
     it('validate_localValid', () => {
         mount(<EmailComponentWithLocalValidation email={testEmailValid} />);
         cy.contains(testEmailValid);
-        cy.get(".invalid").should("not.be.visible");
+        cy.get(".invalid").should("not.exist");
     });
 
     it('validate_localInvalid', () => {
@@ -31,24 +33,49 @@ describe('useValidation', () => {
     it('validate_remoteValid', () => {
         mount(<EmailComponentWithRemoteValidation email={testEmailValid} />);
         cy.wait(2000).contains(testEmailValid);
-        cy.get(".invalid").should("not.be.visible");
+        cy.get(".invalid").should("not.exist");
     });
 
     it('validate_remoteInvalid', () => {
         mount(<EmailComponentWithRemoteValidation email={testEmailInvalid} />);
         cy.wait(2000).get(".invalid").should("be.visible");
     });
+
+    it('validate_localValid_setValue', () => {
+        const Component = () => {
+            const email = useValidation(undefined, isValidEmail);
+
+            useEffect(() => {
+                email.setValue(testEmailValid);
+                if (email.value) {
+                    email.validate(email.value);
+                }
+             }, [email]);
+
+            return (
+                <div>
+                    <input defaultValue={email.value}></input>
+                    <p>{email.value}</p>
+                    {email.error && <p className="invalid">Invalid email</p>}
+                </div>
+            )
+        };
+
+        mount(<Component />);
+        cy.wait(100).contains(testEmailValid);
+        cy.get(".invalid").should("not.exist");
+    });
 });
 
 describe('formUtils', () => {
     it('validateFields_localValid', () => {
         const handleValidateFields = (r) => expect(r).eq(false);
-        mount(<EmailComponentWithFields email={testEmailValid} emailValidation={emailValidation} onValidateFields={handleValidateFields} />)
+        mount(<EmailComponentWithFields email={testEmailValid} emailValidation={isValidEmail} onValidateFields={handleValidateFields} />)
     });
 
     it('validateFields_localInvalid', () => {
         const handleValidateFields = (r) => expect(r).eq(true);
-        mount(<EmailComponentWithFields email={testEmailInvalid} emailValidation={emailValidation} onValidateFields={handleValidateFields} />)
+        mount(<EmailComponentWithFields email={testEmailInvalid} emailValidation={isValidEmail} onValidateFields={handleValidateFields} />)
     });
 
     it('validateFields_remoteValid', () => {
@@ -72,7 +99,7 @@ describe('formUtils', () => {
     it('submitForm_localValid', () => {
         let values = [];
         const handleFormSubmit = (r) => values = r;
-        mount(<EmailComponentWithSubmit email={testEmailValid} emailValidation={emailValidation} onFormSubmit={handleFormSubmit} />)
+        mount(<EmailComponentWithSubmit email={testEmailValid} emailValidation={isValidEmail} onFormSubmit={handleFormSubmit} />)
         cy.wait(100).then(() => {
             expect(values).to.not.be.empty;
         });
@@ -81,7 +108,7 @@ describe('formUtils', () => {
     it('submitForm_localInvalid', () => {
         let values = [];
         const handleFormSubmit = (r) => values = r;
-        mount(<EmailComponentWithSubmit email={testEmailInvalid} emailValidation={emailValidation} onFormSubmit={handleFormSubmit} />)
+        mount(<EmailComponentWithSubmit email={testEmailInvalid} emailValidation={isValidEmail} onFormSubmit={handleFormSubmit} />)
         cy.wait(100).then(() => {
             expect(values).to.be.empty;
         });
@@ -102,6 +129,63 @@ describe('formUtils', () => {
         mount(<EmailComponentWithSubmit email={testEmailInvalid} emailValidation={emailValidationRemote} onFormSubmit={handleFormSubmit} />)
         cy.wait(2000).then(() => {
             expect(values).to.be.empty;
+        });
+    });
+
+    it('submitForm_localValidReturnValues', () => {
+        const expectedResult = { result: 'ok' };
+        let result = null;
+
+        function Component() {
+            const email = useValidation(testEmailValid, isValidEmail);
+            useEffect(() => {
+                result = submitForm([email], () => expectedResult);
+            }, []);
+            return (<div></div>);
+        }
+
+        mount(<Component />);
+
+        cy.then(() => {
+            expect(result).to.eql(expectedResult);
+        });
+    });
+
+    it('submitForm_localInvalidReturnValues', () => {
+        const expectedResult = { result: 'ok' };
+        let result = null;
+
+        function Component() {
+            const email = useValidation(testEmailInvalid, isValidEmail);
+            useEffect(() => {
+                result = submitForm([email], () => expectedResult);
+            }, []);
+            return (<div></div>);
+        }
+
+        mount(<Component />);
+
+        cy.then(() => {
+            expect(result).to.be.undefined;
+        });
+    });
+
+    it('submitForm_remoteValidReturnValues', () => {
+        const expectedResult = { result: 'ok' };
+        let result = null;
+
+        function Component() {
+            const email = useValidation(testEmailValid, emailValidationRemote);
+            useEffect(async () => {
+                result = await submitForm([email], () => expectedResult);
+            }, []);
+            return (<div></div>);
+        }
+
+        mount(<Component />);
+
+        cy.wait(2000).then(() => {
+            expect(result).to.eql(expectedResult);
         });
     });
 });
