@@ -1,9 +1,8 @@
 import React, { useEffect } from 'react';
-import { useValidation } from '../../src/useValidation';
+import { useValidation, isValidEmail } from '../../src';
 
 const EmailComponentWithLocalValidation = (props) => {
-    const emailValidation = (value) => /^[a-zA-Z0-9.!#$%&’*+/=?^_`{|}~-]+@[a-zA-Z0-9-]+(?:\.[a-zA-Z0-9-]+)*$/.test(value);
-    const email = useValidation(props.email, emailValidation);
+    const email = useValidation(props.email, isValidEmail);
 
     useEffect(() => {
        email.validate(email.value);
