@@ -26,7 +26,8 @@ export const useValidation = (defaultValue, validationFn, config) => {
         ...config
     };
 
-    const [value, setValue] = useState(defaultValue);
+    const [resetToValue, setResetToValue] = useState(defaultValue);
+    const [value, setValue] = useState(resetToValue);
     const [error, setError] = useState(false);
     const [dirty, setDirty] = useState(false);
 
@@ -60,6 +61,11 @@ export const useValidation = (defaultValue, validationFn, config) => {
         }
     };
 
+    const _handleSetValue = (v) => {
+        setResetToValue(v);
+        setValue(v);
+    }
+
     const _setValidationResult = (isError, config) => {
         const activeConfig = config ?? _config;
 
@@ -90,10 +96,16 @@ export const useValidation = (defaultValue, validationFn, config) => {
 
     return {
         value,
-        error,
         onChange,
         onBlur,
+        error,
+        setValue: _handleSetValue,
         validate,
-        reset
+        reset,
+        props: {
+            value,
+            onChange,
+            onBlur
+        }
     };
 };
