@@ -11,6 +11,11 @@ export declare interface Field {
     onBlur: (event: any, config: FieldConfig) => void,
     validate: (v: any, config: FieldConfig) => boolean | Promise<boolean>,
     reset: () => void,
+    props: {
+        value: Field['value'],
+        onChange: Field['onChange'],
+        onBlur: Field['onBlur']
+    }
 }
 
 export declare interface Fields {
@@ -21,11 +26,13 @@ export declare function useValidation(initialValue: any, validationFn: (() => bo
 
 export declare function extractValues(fields: Fields): object;
 
+export declare function setValues(fields: Fields, values: object): void;
+
 export declare function validateFields(fields: Fields): boolean | Promise<boolean>;
 
 export declare function resetFields(fields: Fields): void;
 
-export declare function submitForm(fields: Fields, onSubmit: () => object): void;
+export declare function submitForm(fields: Fields, onSubmit: (values: object) => object): Promise<object> | object | undefined;
 
 export declare function cancelForm(fields: Fields, onCancel: () => void): void;
 
